test: Do not add USB disk on RHEL 8 and CS8 because they don't have USB disk attached issue fixed

+++ test/edge-simplified-installer.sh
@@ -36,26 +36,27 @@ SSH_KEY=key/ostree_key
 SSH_KEY_PUB=$(cat "${SSH_KEY}".pub)
 EDGE_USER_PASSWORD=foobar
 
-# Prepare osbuild-composer repository file
-sudo mkdir -p /etc/osbuild-composer/repositories
-
 case "${ID}-${VERSION_ID}" in
     "rhel-8.8")
         OSTREE_REF="rhel/8/${ARCH}/edge"
         OS_VARIANT="rhel8-unknown"
+        USB_INSTALLATION="false"
         ;;
     "rhel-9.2")
         OSTREE_REF="rhel/9/${ARCH}/edge"
         OS_VARIANT="rhel9-unknown"
+        USB_INSTALLATION="true"
         ;;
     "centos-8")
         OSTREE_REF="centos/8/${ARCH}/edge"
         OS_VARIANT="centos-stream8"
+        USB_INSTALLATION="false"
         ;;
     "centos-9")
         OSTREE_REF="centos/9/${ARCH}/edge"
         OS_VARIANT="centos-stream9"
         BOOT_ARGS="uefi,firmware.feature0.name=secure-boot,firmware.feature0.enabled=no"
+        USB_INSTALLATION="true"
         ;;
     *)
         echo "unsupported distro: ${ID}-${VERSION_ID}"
@@ -168,7 +169,9 @@ clean_up () {
     fi
     sudo virsh undefine "${IMAGE_KEY}" --nvram
     sudo virsh vol-delete --pool images "${IMAGE_KEY}.qcow2"
-    sudo virsh vol-delete --pool images "usb.qcow2"
+    if [[ "$USB_INSTALLATION" == "true" ]]; then
+        sudo virsh vol-delete --pool images "usb.qcow2"
+    fi
 
     # Remove simplified installer ISO file
     sudo rm -rf "$ISO_FILENAME"
@@ -349,25 +352,46 @@ sudo restorecon -Rv /var/lib/libvirt/images/
 greenprint "📋 Create libvirt image disk"
 LIBVIRT_IMAGE_PATH=/var/lib/libvirt/images/${IMAGE_KEY}.qcow2
 sudo qemu-img create -f qcow2 "${LIBVIRT_IMAGE_PATH}" 20G
-LIBVIRT_FAKE_USB_PATH=/var/lib/libvirt/images/usb.qcow2
-sudo qemu-img create -f qcow2 "${LIBVIRT_FAKE_USB_PATH}" 16G
 
-greenprint "📋 Install edge vm via simplified installer"
-sudo virt-install --name="${IMAGE_KEY}"\
-                  --disk path="${LIBVIRT_IMAGE_PATH}",format=qcow2 \
-                  --disk path="${LIBVIRT_FAKE_USB_PATH}",format=qcow2 \
-                  --ram 3072 \
-                  --vcpus 2 \
-                  --network network=integration,mac=34:49:22:B0:83:30 \
-                  --os-type linux \
-                  --os-variant "$OS_VARIANT" \
-                  --boot "${BOOT_ARGS}" \
-                  --cdrom "/var/lib/libvirt/images/${ISO_FILENAME}" \
-                  --tpm backend.type=emulator,backend.version=2.0,model=tpm-crb \
-                  --nographics \
-                  --noautoconsole \
-                  --wait=-1 \
-                  --noreboot
+if [[ "$USB_INSTALLATION" == "true" ]]; then
+    # Create a disk to simulate USB device to test USB installation
+    # New growfs service dealing with LVM in simplified installer breaks USB installation
+    LIBVIRT_FAKE_USB_PATH=/var/lib/libvirt/images/usb.qcow2
+    sudo qemu-img create -f qcow2 "${LIBVIRT_FAKE_USB_PATH}" 16G
+
+    greenprint "📋 Install edge vm via simplified installer with USB attached"
+    sudo virt-install --name="${IMAGE_KEY}"\
+                    --disk path="${LIBVIRT_IMAGE_PATH}",format=qcow2 \
+                    --disk path="${LIBVIRT_FAKE_USB_PATH}",format=qcow2 \
+                    --ram 2048 \
+                    --vcpus 2 \
+                    --network network=integration,mac=34:49:22:B0:83:30 \
+                    --os-type linux \
+                    --os-variant "$OS_VARIANT" \
+                    --boot "${BOOT_ARGS}" \
+                    --cdrom "/var/lib/libvirt/images/${ISO_FILENAME}" \
+                    --tpm backend.type=emulator,backend.version=2.0,model=tpm-crb \
+                    --nographics \
+                    --noautoconsole \
+                    --wait=-1 \
+                    --noreboot
+else
+    greenprint "📋 Install edge vm via simplified installer"
+    sudo virt-install --name="${IMAGE_KEY}"\
+                      --disk path="${LIBVIRT_IMAGE_PATH}",format=qcow2 \
+                      --ram 2048 \
+                      --vcpus 2 \
+                      --network network=integration,mac=34:49:22:B0:83:30 \
+                      --os-type linux \
+                      --os-variant "$OS_VARIANT" \
+                      --boot "${BOOT_ARGS}" \
+                      --cdrom "/var/lib/libvirt/images/${ISO_FILENAME}" \
+                      --tpm backend.type=emulator,backend.version=2.0,model=tpm-crb \
+                      --nographics \
+                      --noautoconsole \
+                      --wait=-1 \
+                      --noreboot
+fi
 
 # Start VM.
 greenprint "💻 Start HTTP BOOT VM"
@@ -406,7 +430,7 @@ ansible_become_pass=${EDGE_USER_PASSWORD}
 EOF
 
 # Test IoT/Edge OS
-sudo podman run -v "$(pwd)":/work:z -v "${TEMPDIR}":/tmp:z --rm quay.io/rhel-edge/ansible-runner:latest ansible-playbook -v -i "${TEMPDIR}"/inventory -e os_name=redhat -e ostree_commit="${INSTALL_HASH}" -e ostree_ref="${REF_PREFIX}:${OSTREE_REF}" check-ostree.yaml || RESULTS=0
+sudo podman run -v "$(pwd)":/work:z -v "${TEMPDIR}":/tmp:z --rm quay.io/rhel-edge/ansible-runner:latest ansible-playbook -v -i /tmp/inventory -e os_name=redhat -e ostree_commit="${INSTALL_HASH}" -e ostree_ref="${REF_PREFIX}:${OSTREE_REF}" check-ostree.yaml || RESULTS=0
 check_result
 
 
@@ -537,7 +561,7 @@ ansible_become_pass=${EDGE_USER_PASSWORD}
 EOF
 
 # Test IoT/Edge OS
-sudo podman run -v "$(pwd)":/work:z -v "${TEMPDIR}":/tmp:z --rm quay.io/rhel-edge/ansible-runner:latest ansible-playbook -v -i "${TEMPDIR}"/inventory -e os_name=redhat -e ostree_commit="${UPGRADE_HASH}" -e ostree_ref="${REF_PREFIX}:${OSTREE_REF}" check-ostree.yaml || RESULTS=0
+sudo podman run -v "$(pwd)":/work:z -v "${TEMPDIR}":/tmp:z --rm quay.io/rhel-edge/ansible-runner:latest ansible-playbook -v -i /tmp/inventory -e os_name=redhat -e ostree_commit="${UPGRADE_HASH}" -e ostree_ref="${REF_PREFIX}:${OSTREE_REF}" check-ostree.yaml || RESULTS=0
 check_result
 
 # Final success clean up
